Move media update usecase and remove unnecessary modules
