+++ feature/settings/internal/build.gradle
@@ -21,8 +21,8 @@ dependencies {
 
     implementation project(":library:compose-ext:public")
     implementation project(":library:coroutine-ext:public")
+    implementation project(":library:holoclient:public")
     implementation project(":library:settings:public")
-    implementation project(":usecase")
 
     // Dependency Injection
     implementation "com.google.dagger:hilt-android:$hiltVersion"
