+++ library/holoclient/internal/build.gradle
@@ -16,6 +16,8 @@ dependencies {
     // Project
     api project(":library:holoclient:public")
     implementation project(":api")
+    implementation project(":core")
+    implementation project(":library:settings:public")
 
     // Kotlin
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion"
