+++ library/holoclient/internal/src/main/kotlin/com/minirogue/holoclient/api/GetMediaFromApi.kt
@@ -2,8 +2,6 @@ package com.minirogue.holoclient.api
 
 import android.util.Log
 import com.minirogue.api.media.StarWarsMedia
-import com.minirogue.holoclient.GetMediaFromApi
-import com.minirogue.starwarscanontracker.core.result.HoloResult
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.engine.okhttp.OkHttp
@@ -16,8 +14,9 @@ import java.io.IOException
 import javax.inject.Inject
 
 private const val TAG = "GetMediaFromApiImpl"
-internal class GetMediaFromApiImpl @Inject constructor() : GetMediaFromApi {
-    override suspend fun invoke(): HoloResult<List<StarWarsMedia>> {
+
+internal class GetMediaFromApi @Inject constructor() {
+    suspend operator fun invoke(): HoloResult<List<StarWarsMedia>> {
         return HttpClient(OkHttp) {
             install(ContentNegotiation) {
                 json()
@@ -25,7 +24,7 @@ internal class GetMediaFromApiImpl @Inject constructor() : GetMediaFromApi {
         }.use { client ->
             try {
                 val result: List<StarWarsMedia> =
-                    client.get("https://minirogue.github.io/holocanon-api/media.json").body()
+                        client.get("https://minirogue.github.io/holocanon-api/media.json").body()
                 HoloResult.Success(result)
             } catch (responseException: ResponseException) {
                 Log.i(TAG, "failed response $responseException")
