+++ library/holoclient/internal/src/main/kotlin/com/minirogue/holoclient/api/HoloResult.kt
@@ -1,6 +1,6 @@
-package com.minirogue.starwarscanontracker.core.result
+package com.minirogue.holoclient.api
 
-sealed interface HoloResult<T> {
+internal sealed interface HoloResult<T> {
     data class Success<T>(val value: T) : HoloResult<T>
     data class Failure<T>(val reason: Throwable) : HoloResult<T>
 }
