+++ library/holoclient/internal/src/main/kotlin/com/minirogue/holoclient/usecase/UpdateMediaDatabaseUseCase.kt
@@ -1,4 +1,4 @@
-package com.minirogue.usecase
+package com.minirogue.holoclient.usecase
 
 import android.content.Context
 import android.net.ConnectivityManager
@@ -7,15 +7,15 @@ import android.widget.Toast
 import com.minirogue.api.media.Company
 import com.minirogue.api.media.MediaType
 import com.minirogue.api.media.StarWarsMedia
-import com.minirogue.holoclient.GetApiMediaVersion
-import com.minirogue.holoclient.GetMediaFromApi
+import com.minirogue.holoclient.api.GetApiMediaVersion
+import com.minirogue.holoclient.api.GetMediaFromApi
+import com.minirogue.holoclient.api.HoloResult
 import com.minirogue.starwarscanontracker.core.model.UpdateFilters
 import com.minirogue.starwarscanontracker.core.model.room.MediaDatabase
 import com.minirogue.starwarscanontracker.core.model.room.entity.CompanyDto
 import com.minirogue.starwarscanontracker.core.model.room.entity.MediaItem
 import com.minirogue.starwarscanontracker.core.model.room.entity.MediaNotes
 import com.minirogue.starwarscanontracker.core.model.room.entity.Series
-import com.minirogue.starwarscanontracker.core.result.HoloResult
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -32,10 +32,8 @@ import settings.usecase.GetAllSettings
 import settings.usecase.SetLatestDatabaseVersion
 import javax.inject.Inject
 
-private const val TAG = "UpdateMediaDatabase"
-
 @Suppress("LongParameterList")
-public class UpdateMediaDatabaseUseCase @Inject constructor(
+internal class UpdateMediaDatabaseUseCase @Inject constructor(
         private val updateFilters: UpdateFilters,
         private val connectivityManager: ConnectivityManager,
         private val getApiMediaVersion: GetApiMediaVersion,
@@ -44,51 +42,53 @@ public class UpdateMediaDatabaseUseCase @Inject constructor(
         private val getSettings: GetAllSettings,
         private val setLatestDatabaseVersion: SetLatestDatabaseVersion,
         private val context: Context,
-) {
-    operator fun invoke(forced: Boolean = false) = usecaseScope.launch {
-        mutex.withLock {
-            val settings = getSettings().first()
-            val wifiOnly = settings.syncWifiOnly
-            val latestLocalVersion = settings.latestDatabaseVersion
-            if (wifiOnly && connectivityManager.isActiveNetworkMetered) {
-                return@withLock
-            }
-            val latestRemoteVersion = (getApiMediaVersion() as? HoloResult.Success)?.value?.toLong()
-
-            if (!forced && latestRemoteVersion == latestLocalVersion) {
-                return@withLock
-            }
-
-            val seriesMap: MutableMap<String, Int> =
-                    getSeriesMap().toMutableMap()
-            val companyMap: Map<Company, Int> =
-                    getCompanyMap()
-            val typeMap: Map<MediaType, Int> =
-                    getTypeMap()
-
-            (getMediaFromApi() as? HoloResult.Success)?.value?.let { mediaList ->
-                val daoMedia = database.daoMedia
-                val daoSeries = database.daoSeries
-                mediaList.asFlow()
-                        .map { media ->
-                            val series = media.series
-                            if (seriesMap[series] == null && !series.isNullOrEmpty()) {
-                                val seriesId = daoSeries.insert(Series().apply { title = series }).toInt()
-                                seriesMap[series] = seriesId
-                            }
-                            media.toDTO(seriesMap, typeMap, companyMap)
-                        }
-                        .collect {
-                            val insertSucceeded = daoMedia.insert(it)
-                            if (insertSucceeded == -1L) daoMedia.update(it)
-                            daoMedia.insert(MediaNotes(mediaId = it.id))
-                        }
-                if (latestRemoteVersion != null) {
-                    setLatestDatabaseVersion(latestRemoteVersion)
+) : MaybeUpdateMediaDatabase {
+    override fun invoke(forced: Boolean) {
+        usecaseScope.launch {
+            mutex.withLock {
+                val settings = getSettings().first()
+                val wifiOnly = settings.syncWifiOnly
+                val latestLocalVersion = settings.latestDatabaseVersion
+                if (wifiOnly && connectivityManager.isActiveNetworkMetered) {
+                    return@withLock
                 }
-                updateFilters()
+                val latestRemoteVersion = (getApiMediaVersion() as? HoloResult.Success)?.value?.toLong()
+
+                if (!forced && latestRemoteVersion == latestLocalVersion) {
+                    return@withLock
+                }
+
+                val seriesMap: MutableMap<String, Int> =
+                        getSeriesMap().toMutableMap()
+                val companyMap: Map<Company, Int> =
+                        getCompanyMap()
+                val typeMap: Map<MediaType, Int> =
+                        getTypeMap()
+
+                (getMediaFromApi() as? HoloResult.Success)?.value?.let { mediaList ->
+                    val daoMedia = database.daoMedia
+                    val daoSeries = database.daoSeries
+                    mediaList.asFlow()
+                            .map { media ->
+                                val series = media.series
+                                if (seriesMap[series] == null && !series.isNullOrEmpty()) {
+                                    val seriesId = daoSeries.insert(Series().apply { title = series }).toInt()
+                                    seriesMap[series] = seriesId
+                                }
+                                media.toDTO(seriesMap, typeMap, companyMap)
+                            }
+                            .collect {
+                                val insertSucceeded = daoMedia.insert(it)
+                                if (insertSucceeded == -1L) daoMedia.update(it)
+                                daoMedia.insert(MediaNotes(mediaId = it.id))
+                            }
+                    if (latestRemoteVersion != null) {
+                        setLatestDatabaseVersion(latestRemoteVersion)
+                    }
+                    updateFilters()
+                }
+                withContext(Dispatchers.Main) { Toast.makeText(context, "Database Synced", Toast.LENGTH_SHORT).show() }
             }
-            withContext(Dispatchers.Main) { Toast.makeText(context, "Database Synced", Toast.LENGTH_SHORT).show() }
         }
     }
 
@@ -134,7 +134,9 @@ public class UpdateMediaDatabaseUseCase @Inject constructor(
     private fun String.trimQuotes() = this.trim('\"')
 
     companion object {
-        private val mutex = Mutex()
+        private const val TAG = "UpdateMediaDatabase"
+
         private val usecaseScope = CoroutineScope(Job() + Dispatchers.IO)
+        private val mutex = Mutex()
     }
 }
