+++ library/holoclient/public/build.gradle
@@ -1,6 +1,4 @@
 apply from: "$rootProject.projectDir/base-gradles/kotlin-library.gradle"
 
 dependencies {
-    api project(":api")
-    api project(":library:core:public")
 }
