--- library/holoclient/public/src/main/kotlin/com/minirogue/holoclient/GetApiMediaVersion.kt
@@ -1,7 +1,0 @@
-package com.minirogue.holoclient
-
-import com.minirogue.starwarscanontracker.core.result.HoloResult
-
-interface GetApiMediaVersion {
-    suspend operator fun invoke(): HoloResult<Int>
-}
