--- library/holoclient/public/src/main/kotlin/com/minirogue/holoclient/GetMediaFromApi.kt
@@ -1,8 +1,0 @@
-package com.minirogue.holoclient
-
-import com.minirogue.api.media.StarWarsMedia
-import com.minirogue.starwarscanontracker.core.result.HoloResult
-
-interface GetMediaFromApi {
-    suspend operator fun invoke(): HoloResult<List<StarWarsMedia>>
-}
