+++ library/holoclient/public/src/main/kotlin/com/minirogue/holoclient/usecase/MaybeUpdateMediaDatabase.kt
@@ -1,0 +1,5 @@
+package com.minirogue.holoclient.usecase
+
+interface MaybeUpdateMediaDatabase {
+    operator fun invoke(forced: Boolean = false)
+}
