+++ settings.gradle
@@ -1,11 +1,9 @@
 include ':core'
 include ':app'
-include ':usecase'
 include ':api'
 include ':feature:settings:internal'
 include ':feature:settings:public'
 include ':library:compose-ext:public'
-include ':library:core:public'
 include ':library:coroutine-ext:public'
 include ':library:filters:internal'
 include ':library:filters:public'
